Moved resources to own folder
