+++ mass_api_client/__init__.py
@@ -1,6 +1,6 @@
-from .sample import Sample, DomainSample, IPSample, FileSample, ExecutableBinarySample
-from .report import Report
-from .analysis_request import AnalysisRequest
-from .scheduled_analysis import ScheduledAnalysis
-from .analysis_system import AnalysisSystem
-from .analysis_system_instance import AnalysisSystemInstance
+from .resources.sample import Sample, DomainSample, IPSample, FileSample, ExecutableBinarySample
+from .resources.report import Report
+from .resources.analysis_request import AnalysisRequest
+from .resources.scheduled_analysis import ScheduledAnalysis
+from .resources.analysis_system import AnalysisSystem
+from .resources.analysis_system_instance import AnalysisSystemInstance
