+++ mass_api_client/resources/scheduled_analysis.py
