+++ mass_api_client/schemas/analysis_system_instance.py
@@ -1,5 +1,6 @@
 from marshmallow import fields, validate, post_load
-from mass_api_client.analysis_system_instance import AnalysisSystemInstance
+
+from mass_api_client.resources.analysis_system_instance import AnalysisSystemInstance
 from .base import BaseSchema
